+++ app/src/hid_device.py
@@ -1,12 +1,14 @@
 import os
+import time
 
 from .a_device import ADevice
-from .device_data import DeviceDescriptors
+from .device_data import DeviceDescriptors, KeycodeDict
 from .stdout_writer import StdoutWriter
 
 
 class HID(ADevice):
     '''
+    HID Keyboard
     ### Reference
     + [How to config gadget](https://docs.kernel.org/usb/gadget_configfs.html)
 
@@ -20,9 +22,10 @@ class HID(ADevice):
             uname -r
             ```
     '''
-    def __init__(self, hid_descriptor: DeviceDescriptors, hid_function: str) -> None:
+    def __init__(self, hid_descriptor: DeviceDescriptors, hid_function: str, input_string: str) -> None:
         self.hid_root = self.USB_CONFIGFS_HOME
         self.hid_function = hid_function
+        self.input_string = input_string
         HID.DESCRIPTOR = hid_descriptor
         super().__init__()
 
@@ -45,7 +48,23 @@ class HID(ADevice):
     def enable_the_gadget(self):
         udcname = os.popen("ls /sys/class/udc").read().split("\n")[0] # read udcname
         os.system("sudo bash -c 'echo {} > {}/g1/UDC'".format(udcname, self.hid_root))
+        os.system("sudo chmod 777 /dev/hidg0")  
         StdoutWriter.write("mount job finished!\n")
+        StdoutWriter.write(f"typing '{self.input_string}'\n")
+        self.type_string()
 
     def disable_the_gadget(self):
         return super().disable_the_gadget()
+    
+    def type_string(self):
+        '''
+        Type a string by writing HID reports to /dev/hidg0 using keycodedict
+        '''
+        NULL = b'\x00'
+        with open("/dev/hidg0", "wb") as char_device:
+            time.sleep(1)
+            for char in self.input_string:
+                keycode = KeycodeDict.keycode_dict.get(char.lower())
+                if keycode:
+                    char_device.write(NULL*2 + keycode + NULL*5)  # press
+                    char_device.write(NULL*8)                    # release